IMPALA-14154: IllegalStateException with Iceberg table with DELETE
Planning a query on an Iceberg table runs on IllegalStateException in
the following case:
  - the table has delete files for each data file (i.e. no data file
    without deletes)
  AND
  - there is an anti-join on top of the Iceberg delete operation
    (IcebergDeleteNode or HashJoinNode).

The exception is thrown by a Preconditions check in
SingleNodePlanner.createJoinNode() because there is no null-matching EQ
operator. This happens because the conjunct that should be there is
discarded, ultimately in Analyzer.canEvalAntiJoinedConjunct(). The
reason is that the tuple ids passed to that function include the delete
part of the Iceberg scan, but the conjunct (correctly) only refers to
the tuple of the data files.

Note that this does not happen if we have two regular anti-joins on top
of each other because in that case the getTblRefIds() method of the
bottom anti-join returns a single TableRefId, the one corresponding to
the (inline) view containing it, not the two TableRefIds corresponding
to its two child nodes. The conjunct references this TupleId, so
Analyzer.canEvalAntiJoinedConjunct() returns true. This is not the case
with Iceberg delete operations because there is no (inline) view
involved.

This commit solves the issue by setting the TableRefIds of the node
corresponding to the Iceberg delete operation (IcebergDeleteNode or
HashJoinNode) to only the table ref that corresponds to the data files,
not the delete files.

Testing:
  - added a test in iceberg-v2-read-position-deletes.test that
    reproduces the issue.

Change-Id: If2c03fe3da44dc0516ebdf32430416a1059d37b2
Reviewed-on: http://gerrit.cloudera.org:8080/23051
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>