IMPALA-14161: Thrift SSL tests fail on Centos 7
After the rebase of the Kudu code in to v1.17.1 (see IMPALA-14066), some
backend tests in thrift-server-test.cc stopped working on CentOs 7.9.

This is because the list of available TLS ciphers has been updated in
Kudu commit https://gerrit.cloudera.org/#/c/17268/, which was picked up
in the rebase.

This commit fixes the tests by making sure that the ThriftServers
started in the tests receive the list of ciphers.

Testing:
 - tested the patch on CentOS 7.9.

Change-Id: I2ade7e7308c05baf640a6cbfd62d10f9689a5d9a
Reviewed-on: http://gerrit.cloudera.org:8080/23069
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>