IMPALA-14141 (Addendum): Restore Hive compactor settings after HIVE-28662
This patch restores Hive compactor settings that were changed in
HIVE-28662 but not in IMPALA-14141. Without the addendum, some tests
like testAcidMinorCompactionLoading() could still fail.

Note that this doesn't change the value for
metastore.housekeeping.leader.election that was also changed in
HIVE-28662, as that doesn't control compaction and either values works.

Change-Id: Ifd651aeacf3deba1a8272803207192d5da32a2e4
Reviewed-on: http://gerrit.cloudera.org:8080/23082
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>