IMPALA-14076: Improve readability of workload management query
This patch improve the readability of workload management's insert dml
query profiles by:
1. Add a newline between each entry in the VALUES clause.
2. Remove analyzed query from the PLAN column in both tables.

For second one, a new query option HIDE_ANALYZED_QUERY is added. If this
option is set to True, 'Analyzed query' will not be printed in Plan
section of runtime profile. This is helpful for long SQL such as
workload management's insert dml query.

Testing:
- Add explain test case for HIDE_ANALYZED_QUERY option.
- Manually run some queries in minicluster with enabled workload
  management. Confirmed that both improvement happen in DML runtime
  profile.

Change-Id: I30576795dbc2af27a6879684f3757becfd8fc8d0
Reviewed-on: http://gerrit.cloudera.org:8080/23085
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>