IMPALA-14697: Generate TPC-H/TPC-DS data in parallel
For performance test jobs, generating the datasets at higher
scales can take significant time. Currently, this is running
a single invocation of the generator binary in a single
thread. Both TPC-H and TPC-DS have an ability to generate data
in parallel, producing multiple files for larger tables. This
keeps the current behavior for scale factor 1 to avoid having
a different number of files. This generates higher scale
factors with $(nproc) parallel threads. The LOAD DATA statements
for these tables already reference the directory name itself,
so it doesn't matter that there are more files.

Testing:
 - Ran core job
 - Loaded TPC-H scale 42
 - Loaded TPC-DS scale 20

Change-Id: I7ba13e2275be2ac1a5ae8f9354c947d9f1adf263
Reviewed-on: http://gerrit.cloudera.org:8080/23092
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>