IMPALA-14207: Hook log4j to Glog in AdmissionD
log4j logs does not shows up in dedicated AdmissionD service. This is
because GlogAppender.Install is not called anywhere. It is called in
constructor of JniCatalog.java and JniFrontend.java, but AdmissionD does
not utilize either of them.

AdmissionD has call path to JniRequestPoolService.java. This patch adds
isAdmissiond parameter to JniRequestPoolService constructor to call
GlogAppender.Install and hook up log4j to Glog. Only AdmissionD sets
this parameter to true when initializing JniRequestPoolService. Note
that this hook up only happen if fair_scheduler_allocation_path flag is
not empty.

Testing:
Run few tests in TestAdmissionControllerWithACService and confirm log4j
logs appears in admissiond.INFO

Change-Id: Ie28ec017f70c79d6ffdc56daa5aaeb9efcb8bb79
Reviewed-on: http://gerrit.cloudera.org:8080/23140
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Yida Wu <wydbaggio000@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>