IMPALA-13888: LEFT ANTI JOIN is not working with Iceberg V2 tables on the right side
If an Iceberg table contains delete files, queries where it is on the
right side of a left anti-join fail:
  select *
  from alltypes a
    LEFT ANTI JOIN
    iceberg_v2_positional_update_all_rows b
  ON a.id = b.i;

  AnalysisException: Illegal column/field reference
  'b.input__file__name' of semi-/anti-joined table 'b'

This is because semi-joined tuples need to be made visible explicitly in
order for paths pointing inside them to be resolvable, see
Analyzer::resolvePaths().

This commit adds code to IcebergScanPlanner to make the tuple containing
the virtual fields visible if it is semi-joined.

Testing:
  - Added regressions tets in iceberg-v2-read-position-deletes.test.

Change-Id: I19de9c7c7ed1d61cde281d270c4cc3ce0b7c582d
Reviewed-on: http://gerrit.cloudera.org:8080/23147
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>