IMPALA-14209: Fixed incorrect usage of WriteEventInfos with commitTxnMessage
HIVE-28976 modified CommitTxnMessage to pass write ids in the message.
But, catalogd do not expect write ids in the CommitTxnMessage.
writeEventInfos is explicitly fetched from HMS using
getAllWriteEventInfo() API and populate them to its local
CommitTxnMessage object using CommitTxnMessage.addWriteEventInfo().
Since the write ids are already present in CommitTxnMessage, when we
do addWriteEventInfo() on it, write ids are added again to the
writeIds list in CommitTxnMessage and cause the
IndexOutOfBoundsException upon CommitTxnMessage.getTableObj(i).

Fix is to not populate the fetched WriteEventInfos to local
CommitTxnMessage object. Instead, use them directly.

Testing:
- Executed existing tests.

Change-Id: Ie9b955b2b7427db9771ffec7ac5a16617f5ea022
Reviewed-on: http://gerrit.cloudera.org:8080/23152
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>