IMPALA-14215: Fix tarball creation for extra Python versions
make_shell_tarball.sh previously used the system Python to copy the
required packages to each extra python versions. This caused version
mismatches as the compiled modules are not portable between Python
versions. This change modifies the logic to use the cache of the
provided extra version instead of the system Python.

Tests:
 - manually validated that each provided extra Python contains their
   respective compiled dependencies

Change-Id: Iaee9b3a98b73fd1faf1b7c8ba4b388722add6fb4
Reviewed-on: http://gerrit.cloudera.org:8080/23160
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>