IMPALA-14255: Install Java 17 in bootstrap_build.sh
This patch bumps the Java version installed during bootstrap_build.sh to
Java 17 to keep the precommit environment consistently on Java 17.

bin/bootstrap_build.sh is a simplified setup script used instead of
bootstrap_system.sh when setting up an Impala environment limited to
compilation only. It is used mainly during the initial, lightweight
precommit checks on jenkins.impala.io when a patchset is submitted for
review.

This setup script was not updated with the Java version change from Java
8 to Java 17, so it became out of synch with the general assumption of
building Impala 5.x versions for Java 17.

This patch also removes a special case reserved for Ubuntu 14.04, which
is now not supported by Impala.

Tested automatically by submitting the patch for review.

Change-Id: I796c6004e13aeca536b339fee765f79f39cc2ea1
Reviewed-on: http://gerrit.cloudera.org:8080/23201
Reviewed-by: Jason Fehr <jfehr@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>