IMPALA-14059: deflake test_hms_event_sync_timeout
test_hms_event_sync_timeout adds a sleep in events processing and runs a
SELECT in Impala after an INSERT in Hive. The Impala SELECT statement is
submitted with sync_hms_events_wait_time_s=2 and it's expected that
changes done in Hive haven't been applied in catalogd yet. However, the
changes is applied by a single event (ADD_PARTITION) and the event
processing delay is just 2s which is not longer enough. Sometimes the
event is applied just before the waitForHmsEvent request times out. So
the query still sees the latest results and fails the test.

This increases the event processing delay to 4s to deflake the test.

Change-Id: I91e9cbf234360446422259e274161a01a43ea3d9
Reviewed-on: http://gerrit.cloudera.org:8080/23207
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>